+++ examples/CsQuery.MvcApp/CsQuery.MvcApp/CsQueryView/CsQueryView.cs
@@ -11,8 +11,28 @@ using CsQuery.MvcApp.Controllers;
 namespace CsQuery.Mvc
 {
     /// <summary>
-    /// A custom view that provides access to a CQ object of the html before it is output
+    /// A custom view that provides access to a CQ object of the html before it is output. When used
+    /// as the view engine for a controller marked with ICsQueryController, it will populate the one
+    /// property of that interface "Dom" with a CQ object representing the final HTML for this view,
+    /// and then call for the following methods:
+    /// 
+    /// 
+    /// void Cq_Start()     // called for any action first
+    /// 
+    /// void Cq_Action()    // called for a specfic action
+    /// 
+    /// void Cq_End()       // called for any action last
+    /// 
+    /// 
+    /// void Cq_PartialViewName()          // called whenever the partial view
+    ///                                    // that matches either the  file name
+    /// 
+    /// void Cq_Action_PartialViewName()   // caled whenever the matching partial view is
+    ///                                    // rendered just for this action.
     /// </summary>
+    ///
+    /// <remarks>   James Treworgy, 7/1/2012. </remarks>
+
     public class CsQueryView : RazorView
     {
         #region constructors
@@ -56,11 +76,11 @@ namespace CsQuery.Mvc
         protected override void RenderView(ViewContext viewContext, TextWriter writer, object instance)
         {
 
-            if (viewContext.Controller is CsQueryController)
+            if (viewContext.Controller is ICsQueryController)
             {
                 CsQueryController controller = (CsQueryController)viewContext.Controller;
                 
-                if (HasCqMethods(viewContext) && !IsPartial)
+                if (HasCqMethods(viewContext))
                 {
                     var sb = new StringBuilder();
                     var sw = new StringWriter(sb);
@@ -78,18 +98,24 @@ namespace CsQuery.Mvc
                         // viewName is the name of the cshtml file with underscores e.g "ASP.namespace_subspace_shared__layout_cshtml"
                         // we've mapped things that could be partials with the "_cshtml" attached to the key in the dictionary
 
-                        string viewName = instance.ToString();
-                        
-                        string prefix = controllerName + "_" + currentAction;
+                        var viewInfo = ParseInstance(instance);
+                        string prefix = controllerName + "_Cq_";
+                        string actionPrefix = prefix+currentAction+"_";
 
-                        string key = CqMethods.Keys
-                            .Where(item => viewName.EndsWith(item)).FirstOrDefault();
+                        MethodInfo mi;
 
-                        if (key != null)
+                        if (CqMethods.TryGetValue(prefix+viewInfo.FileName, out mi) ||
+                            CqMethods.TryGetValue(prefix+viewInfo.FilePath, out mi))
                         {
-                            var del = CqMethods[prefix +"_"+ key];
-                            del.Invoke(controller, new object[] { cqDoc  });
-                        } 
+                            mi.Invoke(controller, new object[] { cqDoc });
+                        }
+
+                        // look for an action-specific method
+                        if (CqMethods.TryGetValue(actionPrefix + viewInfo.FileName, out mi) ||
+                            CqMethods.TryGetValue(actionPrefix + viewInfo.FilePath, out mi))
+                        {
+                            mi.Invoke(controller, new object[] { cqDoc });
+                        }
                     }
                     else
                     {
@@ -123,6 +149,35 @@ namespace CsQuery.Mvc
             base.RenderView(viewContext, writer, instance);
         }
 
+        /// <summary>
+        /// Return a structure contiaining the filename only part of a partial view, and the full
+        /// unique name of the view including its relative path concatenated with underscores
+        /// to match a method name.
+        /// </summary>
+        /// <param name="instance"></param>
+        /// <returns></returns>
+        private PartialViewInfo ParseInstance(object instance)
+        {
+            string[] viewPath = ((System.Web.Mvc.WebViewPage)instance).VirtualPath
+                .Split('/')
+                .Skip(1)
+                .ToArray();
+
+            string partialFileName = viewPath.Last();
+            if (partialFileName.IndexOf(".") >= 0)
+            {
+                partialFileName = partialFileName.Split('.')[0];
+                viewPath[viewPath.Length - 1] = partialFileName;
+            }
+
+
+            return new PartialViewInfo
+            {
+                FileName = partialFileName,
+                FilePath = String.Join("_", viewPath)
+            };
+
+        }
 
         private bool HasCqMethods(ViewContext viewContext)
         {
@@ -146,19 +201,13 @@ namespace CsQuery.Mvc
         {
             bool hasMethods = false;
 
-            foreach (var mi in controllerType.GetMethods(BindingFlags.Public | BindingFlags.Instance))
+            foreach (var mi in controllerType.GetMethods(BindingFlags.Public | BindingFlags.Instance | BindingFlags.Static))
             {
                 if (HasCqSignature(mi))
                 {
 
                     string name = mi.Name;
-                    if (name.IndexOf("_",3) > 0)
-                    {
-                        if (!name.EndsWith("_cshtml"))
-                        {
-                            name += "_cshtml";
-                        }
-                    }
+
                     if (hasMethods==false) {
                         hasMethods = true;
                         if (CqMethods == null)
@@ -172,28 +221,59 @@ namespace CsQuery.Mvc
             HasCqMethodsRef[controllerName] = hasMethods;
         }
 
+        /// <summary>
+        /// Return true if the reflected method appears to be a CsQuery target method
+        /// </summary>
+        /// <param name="mi"></param>
+        /// <returns></returns>
         private bool HasCqSignature(MethodInfo mi)
         {
             var pi = mi.GetParameters();
-            //return pi.Length == 1 && pi[0].ParameterType == typeof(CQ);
-            //if (mi.Name.StartsWith("Cq_"))
-            //{
-            //    Console.Write("");
-            //}
-            return mi.Name.StartsWith("Cq_") && 
-                mi.ReturnType == typeof(void) &&
-                (
-                    pi.Length == 0  && mi.Name.IndexOf("_",3)<0 ||
-                    pi.Length == 1 && pi[0].ParameterType == typeof(CQ)
-                );
+
+            if (mi.Name.StartsWith("Cq_"))
+            {
+                if (mi.IsStatic)
+                {
+                    throw new ArgumentException(IncorrectSignatureError());
+                }
+
+                switch (pi.Length)
+                {
+                    case 0:
+                        return true;
+                    case 1:
+                        // verify parameters are correct
+                        if (pi[0].ParameterType != typeof(CQ) ||
+                            mi.ReturnType != typeof(void))
+                        {
+                            throw new ArgumentException(IncorrectSignatureError());
+                        }
+                        return true;
+                    default:
+                        throw new ArgumentException(IncorrectSignatureError());
+                }
+            }
+            else
+            {
+                return false;
+            }
         }
 
+        private string IncorrectSignatureError() {
+            return "Incorrect signature: a method starting with Cq_ must return void, have zero parameters, or a single CQ parameter.";
+        }
         private string GetCurrentAction(ViewContext context)
         {
             return context.Controller.ControllerContext.RouteData.Values
                 .Where(item => item.Key == "action")
                 .First().Value.ToString();
 
+        }
+
+        private class PartialViewInfo
+        {
+            public string FileName;
+            public string FilePath;
         }
     }
 }